pyproject.toml
[build-system]
build-backend = "hatchling.build"
requires = [ "hatchling" ]

[project]
name = "django-googler"
version = "0.0.3"
description = "Django Googler is a simple way to integrate Google Auth Platform with your Django project."
readme = "README.md"
keywords = [ "auth", "django", "djangorestframework", "google", "oauth", "oauth2", "oauthlib" ]
authors = [
  { name = "Justin Mitchel", email = "justin@codingforentrepreneurs.com" },
]
requires-python = ">=3.12"
classifiers = [
  "Development Status :: 4 - Beta",
  "Framework :: Django",
  "Framework :: Django :: 5.2",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
]
dependencies = [
  "django>=5.2",
  "djangorestframework>=3.15",
  "google-auth>=2.41",
  "google-auth-oauthlib>=1.2",
]

urls.Changelog = "https://github.com/jmitchel3/django-googler"
urls.Documentation = "https://github.com/jmitchel3/django-googler"
urls.Funding = "https://github.com/jmitchel3/django-googler"
urls.Repository = "https://github.com/jmitchel3/django-googler"

[dependency-groups]
dev = [
  "pip>=25.2",
  "pre-commit>=4.3",
  "rav>=0.1.1",
]
test = [
  "coverage[toml]>=7.6",
  "pytest>=8.3",
  "pytest-django>=4.9",
]

[tool.black]
line-length = 88
target-version = [ 'py312' ]

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
pythonpath = [ "src", "." ]
testpaths = [ "tests" ]
addopts = "--reuse-db"

[tool.coverage.run]
branch = true
parallel = true
source_pkgs = [ "django_googler" ]
omit = [ "*/tests/*" ]

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.tox]
legacy_tox_ini = """
[tox]
env_list =
    py312
    py313
    py314

[testenv]
package = wheel
wheel_build_env = .pkg
deps =
    coverage[toml]>=7.6
    pytest>=8.3
    pytest-django>=4.9
commands =
    coverage run --parallel-mode -m pytest {posargs}
    coverage combine
    coverage report
"""
